projetonuvem/app/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
requests
python-dotenv
jose
python-multipart
bcrypt